rename grid mod to grid_traits
