+++ datatypes/src/raster/grid_traits.rs
